refactor: Revised structure
