+++ backend-api/src/modules/types.d.ts
@@ -1,7 +1,7 @@
 import type { EntityItem } from 'electrodb';
 
 import type { createLoaders } from '@/backend-api/src/modules/loaders';
-import type { logger } from '@/backend-api/src/lib/logger';
+import type { logger } from '@/backend-api/src/logger';
 
 export type {
   BlogBrandingItem,
