+++ backend-api/src/trpc/context.ts
@@ -3,7 +3,7 @@ import type { CreateAWSLambdaContextOptions } from '@trpc/server/adapters/aws-la
 import type { APIGatewayProxyEventV2 } from 'aws-lambda';
 
 import * as models from '@/backend-api/src/modules/models';
-import { logger } from '@/backend-api/src/lib/logger';
+import { logger } from '@/backend-api/src/logger';
 import { getHeader, parseAuthHeader } from '@/backend-api/src/helpers/api-gateway';
 import { verifyUserToken } from '@/backend-api/src/modules/authentication/tokens';
 import { createLoaders } from '@/backend-api/src/modules/loaders';
