+++ backend-api/src/trpc/index.ts
@@ -1,4 +1,4 @@
-import { router } from '@/backend-api/src/lib/trpc';
+import { router } from '@/backend-api/src/trpc/core';
 
 import { authRouter } from './auth';
 import { quotesQuery } from './quotes';
